LayoutTests/inspector/styles/styles-add-new-rule-expected.txt

Tests that adding a new rule works after switching nodes.

Text
After adding new rule (inspected):
[expanded] element.style { ()
font-size: 12px;

======== Matched CSS Rules ========
[expanded] #inspected { (via inspector)
color: maroon;

[expanded] div { (user agent stylesheet)
display: block;


After adding new rule (other):
[expanded] element.style { ()
color: red;

======== Matched CSS Rules ========
[expanded] #other { (via inspector)

[expanded] div { (user agent stylesheet)
display: block;
